Move storage class to data client
