+++ libs/newsletters-data-client/src/index.ts
@@ -5,3 +5,4 @@ export * from './lib/api-response-type';
 export * from './lib/wizard-button-type';
 export * from './lib/testQuestionaireSchema';
 export * from './lib/emailEmbedSchema';
+export * from './lib/storage';
