+++ libs/newsletters-data-client/src/lib/storage/index.ts
@@ -1,0 +1,2 @@
+export * from './DraftStorage';
+export * from './InMemoryDraftStorage';
