+++ libs/state-machine/src/lib/stateMachineButtonPressed.ts
@@ -1,5 +1,6 @@
+import type { DraftStorage } from '@newsletters-nx/newsletters-data-client';
 import { getFormSchema } from '../schemas';
-import type { StorageInterface, WizardLayout, WizardStepData } from './types';
+import type { WizardLayout, WizardStepData } from './types';
 
 export function setupInitialState(): WizardStepData {
 	return {
@@ -20,7 +21,7 @@ export async function stateMachineButtonPressed(
 	buttonPressed: string,
 	incomingStepData: WizardStepData,
 	wizardLayout: WizardLayout,
-	storageInstance: StorageInterface,
+	storageInstance: DraftStorage,
 ): Promise<WizardStepData> {
 	const currentStepLayout = wizardLayout[incomingStepData.currentStepId];
 	const buttonPressedDetails = currentStepLayout?.buttons[buttonPressed];
